+++ app/assets/stylesheets/thanks.scss
@@ -5,6 +5,8 @@
 *{
 
   font-family: Noto Sans CJK JP,Poppins;
+  box-sizing: border-box;
+
   
 
   a:hover{
@@ -15,4 +17,11 @@
 ::placeholder{
   color:#AAA;
   padding-left:1em;
+}
+
+input:-webkit-autofill,
+input:-webkit-autofill:hover,
+input:-webkit-autofill:focus,
+input:-webkit-autofill:active {
+    transition: background-color 5000s ease-in-out 0s;
 }
