+++ app/views/layouts/application.html.erb
@@ -4,10 +4,9 @@
     <title>ThanksApp</title>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.css">
     <%= stylesheet_link_tag    'application', media: 'all' %>
     <%= javascript_include_tag 'application'%>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.css">
-    <%# , 'data-turbolinks-track': 'reload'  %>
   </head>
   <body>
     <%= yield %>
